pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=77",
  "setuptools-scm[toml]>=8",
]

[project]
name = "pydeps2env"
description = "A python helper to generate conda environment files from project dependencies."
readme = { file = "README.md", content-type = "text/markdown" }
keywords = [
  "conda",
  "pyproject",
]
license = "BSD-3-Clause"
authors = [
  { name = "Çağtay Fabry", email = "cagtay.fabry@bam.de" },
]
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [
  "version",
]
dependencies = [
  "packaging",
  "pyyaml",
  "tomli; python_version<'3.11'",
]
optional-dependencies.test = [
  # needed to run the test suite
  "pytest",
  "pytest-cov",
]
urls.bug_tracker = "https://github.com/CagtayFabry/pydeps2env/-/issues"
urls.repository = "https://github.com/CagtayFabry/pydeps2env"
scripts.pydeps2env = "pydeps2env.generate_environment:main"

[tool.setuptools.packages]
find = { exclude = [
  "test",
  "resources",
] }

[tool.setuptools_scm]
# enable setuptools-scm versioning

[tool.pytest.ini_options]
addopts = "--tb=short --color=yes -rsw --cov=pydeps2env"
testpaths = [
  "test",
]
